requirements.txt
python-telegram-bot==13.8.1
redis==3.5.3
vk-api==11.9.6